LPS-69590 Move
